# Test each benchmark case in own #[test]