--- a/frame/bags-list/src/benchmarks.rs
+++ b/frame/bags-list/src/benchmarks.rs
@@ -134,11 +134,10 @@ frame_benchmarking::benchmarks! {
 			]
 		);
 	}
-}
 
-use frame_benchmarking::impl_benchmark_test_suite;
-impl_benchmark_test_suite!(
-	Pallet,
-	crate::mock::ExtBuilder::default().build(),
-	crate::mock::Runtime,
-);
+	impl_benchmark_test_suite!(
+		Pallet,
+		crate::mock::ExtBuilder::default().build(),
+		crate::mock::Runtime,
+	)
+}