--- a/frame/democracy/src/lib.rs
+++ b/frame/democracy/src/lib.rs
@@ -149,7 +149,7 @@
 //! - `cancel_queued` - Cancels a proposal that is queued for enactment.
 //! - `clear_public_proposal` - Removes all public proposals.
 
-#![recursion_limit = "128"]
+#![recursion_limit = "256"]
 #![cfg_attr(not(feature = "std"), no_std)]
 
 use codec::{Decode, Encode, Input};