Updates the `impl_benchmark_test_suite!` and extends the `benchmarks!` macros.  
Closes #9110

### Syntax extension
It is now possible to use `impl_benchmark_test_suite!` **within** the `benchmarks!` macro.  

Example:  
```rust
benchmarks! {
	dummy {
		…
	}: {
		…
	}

	impl_benchmark_test_suite!(Pallet, crate::tests::new_test_ext(), crate::tests::Test) // ; allowed here
}

```
### Behavioural change
When `impl_benchmark_test_suite!` is used in a `benchmarks!` call, it now generates one `#[test]` function for each bench case.

Old output:   
```pre
test benchmarking::benchmark_tests::test_benchmarks ... ok
```
New output:
```pre
test benchmarking::bench_sort_vector ... ok
test benchmarking::bench_accumulate_dummy ... ok
test benchmarking::bench_set_dummy_benchmark ... ok
```

👉 `benchmarks!` and `impl_benchmark_test_suite!` stay backwards compatible.

### Parallel benchmarks

Since each bench case has its own `#[test]` function, it is possible to run them in parallel.  
A good example is the *contracts* pallet:  

Result with old syntax (after compilation):
```sh
cargo test --features=runtime-benchmarks -- --test-threads 16  58,85s user 0,61s system 108% cpu 54,619 total
```

Result with new syntax (after compilation):
```sh
cargo test --features=runtime-benchmarks -- --test-threads 16  65,28s user 0,87s system 264% cpu 24,981 total
```

### Open points

- [x] I am not sure if I ran all test. Some down-stream tests would be good for the compatibility check.
- [ ] Comments need to be checked for changes.
- [ ] I have not bumped the version.

✄ -----------------------------------------------------------------------------

Thank you for your Pull Request! 🙏

Before you submit, please check that:

- [x] **Description:** You added a brief description of the PR, e.g.:
  - What does it do?
  - What important points should reviewers know?
  - Is there something left for follow-up PRs?
- [x] **Labels:** You labeled the PR appropriately if you have permissions to do so:
  - [ ] `A*` for PR status (**one required**)
  - [ ] `B*` for changelog (**one required**)
  - [ ] `C*` for release notes (**exactly one required**)
  - [ ] `D*` for various implications/requirements
  - [ ] Github project assignment
- [x] **Related Issues:** You mentioned a related issue if this PR is related to it, e.g. `Fixes #228` or `Related #1337`.
- [x] **2 Reviewers:** You asked at least two reviewers to review. If you aren't sure, start with GH suggestions.
- [x] **Style Guide:** Your PR adheres to [the style guide](https://github.com/paritytech/substrate/blob/master/docs/STYLE_GUIDE.md)
  - In particular, mind the maximal line length of 100 (120 in exceptional circumstances).
  - There is no commented code checked in unless necessary.
  - Any panickers in the runtime have a proof or were removed.
- [ ] **Runtime Version:** You bumped the runtime version if there are breaking changes in the **runtime**.
- [ ] **Docs:** You updated any rustdocs which may need to change.

Refer to [the contributing guide](https://github.com/paritytech/substrate/blob/master/docs/CONTRIBUTING.adoc) for details.

✄ -----------------------------------------------------------------------------