@thiolliere 2021-09-29
> > looks good to me once doc about new feature is added
> 
> So how does this work? Where do you want it documented? @thiolliere

I just mean this https://github.com/paritytech/substrate/pull/9860#discussion_r718897986 benchmark have a new feature: improved test suite when impl_benchmark_test_suite is written inside the macro. So that needs to be documented in the rust-doc of the item `benchmark`: here https://github.com/paritytech/substrate/blob/0d44d83e619f6c45597c11b37f3e36e942d22cc2/frame/benchmarking/src/lib.rs#L58-L191