@KiChjang 2021-10-01
Sorry I was late into the game, but I think a follow-up to this would be to figure out a nicer syntax for the `impl_benchmark_test_suite` macro. Since parsing and expanding the macro falls within the responsibility of `benchmarks!`, we don't actually need to keep macro syntax for `impl_benchmark_test_suite` anymore, and can think of a more intuitive one or even a less verbose one for it.