@thiolliere 2021-10-06
we could also make it an attribute macro so that people can have the automatic formatting